% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustTree.R
\name{clustEM}
\alias{clustEM}
\title{Cluster phylogenetic tree using EM}
\usage{
clustEM(k, file = "", text = NULL, label = NULL, maxDim = NULL)
}
\arguments{
\item{k}{A positive integer greater than 1 indicating the number of clusters.}

\item{file}{A character string indicating the path to the newick tree.
Default value is "".}

\item{text}{A character string of the tree in newick format. Default value
is NULL. By default, this is ignored. When this argument is assigned a
value, the argument file is ignored.}

\item{label}{An integer vector of length equal to the number of leaves.
label should contain integers between 0 and k, with indices of 0s
corresponding to indices of leaves that are to be assigned clusters.
label indices of 1 to k correspond to indices of leaves with pre-labeled
clusters. Default value is NULL, meaning no pre-labeled leaves.}

\item{maxDim}{A positive integer that is at least 1, indicating the max
dimension of the coordinates that represent tree leaves. The dimensions
will be less than or equal to the number of leaves. The default value of
maxDim is NULL, for which the full dimension will be used. maxDim can be
set smaller to decrease runtime of EM at the cost of discarding the
least significant dimensions beyond maxDim.}
}
\value{

}
\description{
A function that calculates optimal tree clustering scheme using EM method
given the number of clusters, file directory for a phylogenetic tree or a
newick tree character string, and an optional set of labeled leaves for
semi-supervised clustering.
}
\examples{

}
\references{

}
