% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustTree.R
\name{clustPAM}
\alias{clustPAM}
\title{Cluster phylogenetic tree using PAM}
\usage{
clustPAM(k, file = "", text = NULL)
}
\arguments{
\item{k}{A positive integer greater than 1 indicating the number of clusters.}

\item{file}{A character string indicating the path to the newick tree.
Default value is "".}

\item{text}{A character string of the tree in newick format. Default value
is NULL. By default, this is ignored. When this argument is assigned a
value, the argument file is ignored.}
}
\value{
Returns an S3 object of class PAMclusts with results.
\itemize{
\item distM - A distance matrix of tree leaves.
\item phyloTree - An S3 object of class phylo, containing tree
information.
\item clustering - A positive integer vector indicating the clusters
assigned to each leaf.
\item medoids - A character vector indicating names of leaves that have
been assigned as cluster centers.
\item stats - A matrix containing parameters for each cluster. The
columns are "size", "max_diss", "av_diss", "diameter", and "separation".
Each row represents a computed cluster.
\itemize{
\item size - number of leaves in the cluster.
\item max_diss - maximum dissimilarity between leaves in the cluster
and center of the cluster.
\item av_diss - average dissimilarity between leaves in the cluster
and center of the cluster.
\item diameter - maximum dissimilarity between two leaves in the
cluster.
\item separation - minimum dissimilarity between one leaf of this
cluster and one leaf of another cluster.
}
}
}
\description{
A function that calculates optimal tree clustering scheme using PAM method
given the number of clusters, file directory for a phylogenetic tree or a
newick tree character string.
}
\examples{

}
\references{

}
