% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustCompare.R
\name{comparePAM}
\alias{comparePAM}
\title{Compare different PAM clusterings of the same phylogenetic tree using gap
statistics.}
\usage{
comparePAM(distM, k.max, clusterings = NULL, B = 100L)
}
\arguments{
\item{distM}{A distance matrix of tree leaves.}

\item{k.max}{An integer that is at least 2, indicating the max number of
clusters desired for the tree.}

\item{clusterings}{A list of k.max - 1 positive integer vectors. Each
positive integer vector is usually the clustering output of clustPAM, in
which each leaf node is associated with an integer representing the
cluster it is assigned. Each positive integer vector must be the same
length as the number of leaves in the tree. Clustering vectors must
indicate the clustering scheme of the tree from 2 clusters (k = 2) to
k.max clusters (k = k.max). Clustering vectors must be ordered in the
list by increasing number of clusters. Default value of clusterings is
NULL, in which case PAM clustering will be performed automatically on
distM over 2 <= k <= k.max.}

\item{B}{An integer indicating the number of bootstrap samples. Default
value is 100.}
}
\value{

}
\description{
A function that compares a set of clustering schemes by PAM and computes the
gap statistics from a given range of k.
}
\examples{

}
\references{

}
