% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustTree.R
\name{getPhyloTree}
\alias{getPhyloTree}
\title{Gets a phylo tree object from a tree string or file.}
\usage{
getPhyloTree(file = "", text = NULL)
}
\arguments{
\item{file}{A character string indicating the path to the newick tree.
Default value is "".}

\item{text}{A character string of the tree in newick format. Default value
is NULL. By default, this is ignored. When this argument is assigned a
value, the argument file is ignored.}
}
\value{
Returns an S3 object of class phylo.
}
\description{
A function that checks the input string or file and returns a phylo tree
object. This is a helper function and is not exported.
}
