% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NwkTree}
\alias{NwkTree}
\title{An example of a small Newick tree}
\format{
A character string
}
\source{
\url{https://evolution.genetics.washington.edu/phylip/newicktree.html}
}
\usage{
NwkTree
}
\description{
A tree in Newick format from an estimate of evolutionary relationships
between a selection of eutherian mammals.
}
\examples{
NwkTree
}
\keyword{datasets}
