% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NwkTree2}
\alias{NwkTree2}
\title{An example of a bigger Newick tree}
\format{
A character string
}
\source{
\url{https://academic.oup.com/sysbio/article-abstract/35/2/230/1649654}
}
\usage{
NwkTree2
}
\description{
A made-up tree in Newick format
}
\examples{
NwkTree2
}
\keyword{datasets}
