% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustComparePlot.R
\name{plotGapStat}
\alias{plotGapStat}
\title{Plot gap statistics.}
\usage{
plotGapStat(gapStat, method = "Tibs2001SEmax", color = "steelblue")
}
\arguments{
\item{gapStat}{An object of class "clusGap" indicating the gap statistics to
be plotted. This is usually output from comparePAM, compareEM, or
cluster::clusGap.}

\item{method}{A character string indicating the method used to determine
best k. Default value is "Tibs2001SEmax". Must be one of "globalmax",
"firstmax", "Tibs2001SEmax", "firstSEmax". Tibs2001SEmax is the
method proposed by Tibshirani et al (2001): "the smallest k such that
gap(k) >= gap(k+1) - s_k+1".}

\item{color}{A character string indicating the color of line graph. Default
value is "steelblue".}
}
\value{

}
\description{
A function that plots given gap statistics and finds the best k value
using a user-specified method.
}
\examples{

}
\references{

}
