% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runClustPhy.R
\name{runClustPhy}
\alias{runClustPhy}
\title{Launch Shiny App for ClustPhy}
\usage{
runClustPhy()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for ClustPhy.
The shiny app allows user to perform phylogenetic clustering via an
interactive platform.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
ClustPhy::runClustPhy()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
\author{
{Yuzi Li, \email{rainal.li@mail.utoronto.ca}}
}
