% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustTree.R
\name{clustEM}
\alias{clustEM}
\title{Cluster phylogenetic tree using EM}
\usage{
clustEM(k, file = "", text = NULL)
}
\arguments{
\item{k}{A positive integer greater than 1 indicating the number of clusters.}

\item{file}{A character string indicating the path to the newick tree.
Default value is "".}

\item{text}{A character string of the tree in newick format. Default value
is NULL. By default, this is ignored. When this argument is assigned a
value, the argument file is ignored.}
}
\value{
Returns an S3 object of class EMclusts with results.
\itemize{
\item distM - A distance matrix of tree leaves.
\item phyloTree - An S3 object of class phylo, containing tree
information.
\item clustering - A positive integer vector indicating the clusters
assigned to each leaf.
\item mean - A numeric matrix with columns corresponding to mean
coordinates of cluster centers.
\item bic - A numeric indicating BIC value of the optimal model.
\item model - A character string describing the optimal model used.
}
}
\description{
A function that calculates optimal tree clustering scheme using EM method
given the number of clusters, file directory for a phylogenetic tree or a
newick tree character string, and an optional set of labeled leaves for
semi-supervised clustering.
}
\examples{
# Make 6 clusters from a newick tree using EM
em <- clustEM(6, text = NwkTree2)

}
\references{
Kaufman, L., & Rousseeuw, P. J. (2005). Finding groups in data:
An introduction to cluster analysis. Wiley.
}
